Move code to common (preparation for JDBC batch execution)
changelog_begin
changelog_end
