+++ ledger/participant-integration-api/src/main/scala/platform/store/backend/postgresql/PGTable.scala
@@ -5,15 +5,7 @@ package com.daml.platform.store.backend.postgresql
 
 import java.sql.Connection
 
-trait Table[FROM] {
-  def prepareData(in: Vector[FROM]): Array[Array[_]]
-  def executeUpdate: Array[Array[_]] => Connection => Unit
-}
-
-abstract class BaseTable[FROM](fields: Seq[(String, Field[FROM, _, _])]) extends Table[FROM] {
-  override def prepareData(in: Vector[FROM]): Array[Array[_]] =
-    fields.view.map(_._2.toArray(in)).toArray
-}
+import com.daml.platform.store.backend.common.{BaseTable, Field, Table}
 
 object PGTable {
